requirements.txt
python-telegram-bot[job-queue]==21.*
yt-dlp
httpx
python-dotenv
py7zr